Update chain tokens screen UI to match Figma
Closes #139
